Align with template
