+++ assets/__demo_variables.css
@@ -1,0 +1,46 @@
+/*
+Copyright 2024 D-Wave
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+/*
+Variable declarations for all CSS files.
+Some additional variables are generated automatically in app.py
+and stored in __generated_theme.css.
+
+Dash reads all css files contained in `/assets/` so no imports are necessary.
+*/
+
+:root {
+    --blue-bright: #03B8FF;
+    --blue-light: #2A7DE1;
+    --blue-dark: #074C91;
+    --red-light: #C70039;
+    --red-dark: #900C3F;
+    --orange: #FF7006;
+    --teal-light: #06ECDC;
+    --teal-dark: #17BEBB;
+    --teal-darker: #008C82;
+    --grey-lighter: #EEEEEE;
+    --grey-light: #DDDDDD;
+    --grey: #AAAAAA;
+    --grey-medium: #777777;
+    --grey-dark: #222222;
+    --font: "Helvetica Neue", sans-serif;
+    --banner-height: 5.5rem;
+    --left-col-width: 36rem;
+    --problem-details-height: 12rem;
+    /*** Add new variables here ***/
+    --error-col-width: 42rem;
+}
