+++ assets/_buttons.css
@@ -1,0 +1,46 @@
+/*
+Copyright 2024 D-Wave
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+/* Style rules for buttons */
+
+button {
+    font-size: 1.4rem;
+    line-height: 1.4rem;
+    padding: 1.8rem 3.6rem;
+    height: auto;
+    color: white;
+    transition: all 0.2s ease-in-out;
+    border: none;
+    background-color: var(--theme);
+}
+
+button:hover {
+    filter: brightness(80%);
+    color: white;
+}
+
+#run-button,
+#cancel-button {
+    width: 100%;
+}
+
+#cancel-button {
+    background-color: var(--red-light);
+}
+
+#cancel-button:hover {
+    background-color: var(--red-dark);
+}
