+++ assets/_slider.css
@@ -1,0 +1,98 @@
+/*
+Copyright 2024 D-Wave
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+/* Style rules for the dcc.Slider element */
+
+.rc-slider-tooltip-inner {
+    font-family: var(--font);
+}
+
+.slider {
+    padding-left: 6px !important;
+    padding-right: 8px !important;
+}
+
+.rc-slider-handle,
+.rc-slider-dot-active {
+    border-color: var(--theme-secondary);
+    transition: filter 0.1s ease-in-out;
+}
+
+.rc-slider-handle:active {
+    box-shadow: 0 0 5px var(--theme-secondary);
+}
+
+.rc-slider-handle:active,
+.rc-slider-handle:hover,
+.rc-slider-handle-click-focused,
+.rc-slider-handle-click-focused:focus {
+    border-color: var(--theme-secondary);
+    filter: brightness(80%)
+}
+
+.rc-slider-track {
+    background-color: var(--theme-secondary);
+}
+
+.rc-slider-tooltip-inner {
+    background-color: white;
+    color: var(--grey-dark);
+    border: 1px solid var(--grey);
+    border-radius: 2px;
+}
+
+.rc-slider-tooltip {
+    pointer-events: none;
+}
+
+.rc-slider-tooltip .rc-slider-tooltip-arrow {
+    display: none;
+}
+
+.rc-slider-tooltip-inner:before,
+.rc-slider-tooltip-inner:after {
+    content: '';
+    display: block;
+    position: absolute;
+    width: 0;
+    height: 0;
+}
+
+.rc-slider-tooltip-inner:before {
+    left: 6px;
+    top: -11px;
+    border: 6px solid transparent;
+    border-bottom-color: var(--grey);
+}
+
+.rc-slider-tooltip-inner:after {
+    left: 7px;
+    top: -9px;
+    border: 5px solid transparent;
+    border-bottom-color: white;
+}
+
+.rc-slider-tooltip-placement-top .rc-slider-tooltip-inner:before {
+    bottom: -3px;
+    top: auto;
+    transform: rotate(180deg);
+}
+
+.rc-slider-tooltip-placement-top .rc-slider-tooltip-inner:after {
+    bottom: -1px;
+    top: auto;
+    transform: rotate(180deg);
+}
